requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
cryptography==41.0.7
requests==2.31.0
python-multipart==0.0.6
pymysql==1.1.0
